chore(skygrid): use default value for itemPayload due to #aa4b92b49a53ce23592217f0335654f8dafcf595
