--- apps/skygrid/src/app/skygrid-data2.ts
@@ -1,1530 +1,0 @@
-import { MatDatagridCollapseComponent, setItemPayload } from '@rescoped/components/datagrid';
-
-export const flightStatus = [
-  { status: 'On Time' },
-  { status: 'Delayed' },
-  { status: 'Cancelled' },
-  { status: 'Boarding' },
-  { status: 'Landed' },
-  { status: 'In Flight' },
-];
-
-export const destinations = [
-  { destination: 'Hamburg' },
-  { destination: 'Frankfurt' },
-  { destination: 'Berlin' },
-  { destination: 'Istanbul' },
-  { destination: 'Brussel' },
-  { destination: 'Rome' },
-  { destination: 'Oslo' },
-  { destination: 'Izmir' },
-  { destination: 'Kos' },
-  { destination: 'Madrid' },
-  { destination: 'Helsinki' },
-  { destination: 'London' },
-  { destination: 'Dublin' },
-  { destination: 'Danzig' },
-  { destination: 'San Diego' },
-  { destination: 'Dallas' },
-  { destination: 'San Jose' },
-  { destination: 'Austin' },
-];
-
-export const gates = [
-  { gate: 'A01' },
-  { gate: 'A02' },
-  { gate: 'A03' },
-  { gate: 'A04' },
-  { gate: 'A05' },
-  { gate: 'B01' },
-  { gate: 'B02' },
-  { gate: 'B03' },
-  { gate: 'B04' },
-  { gate: 'C01' },
-  { gate: 'C02' },
-  { gate: 'C03' },
-  { gate: 'C04' },
-];
-
-export interface SkyGridData {
-  collapse: string;
-  date: string;
-  airline: string;
-  flightNumber: string;
-  status: string;
-  origin: string;
-  destination: string;
-  duration: string;
-  estArrival: string;
-  estDeparture: string;
-  terminal: string;
-  gate: string;
-}
-
-const group1 = setItemPayload(<SkyGridData>{ date: '15.10.2022' });
-const group2 = setItemPayload(<SkyGridData>{ date: '16.10.2022' });
-const group3 = setItemPayload(<SkyGridData>{ date: '17.10.2022' });
-
-export const skyGridData: SkyGridData[] = [
-  // row-global
-  setItemPayload(<SkyGridData>{}, { actionType: 'row-global' }),
-
-  // row-group
-  // 15.10.2022
-  setItemPayload(group1, {
-    actionType: 'row-group',
-    parent: true,
-    groupId: 1,
-    rules: {
-      overrides: {
-        collapse: {
-          action: {
-            componentType: MatDatagridCollapseComponent,
-            componentPosition: 'override',
-            data: {
-              item: group1,
-            },
-          },
-        },
-      },
-    },
-  }),
-
-  // row-single
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-2042',
-      status: 'Boarding',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-2931',
-      status: 'Boarding',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1274',
-      status: '',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-8532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-12',
-      status: 'Delayed',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'Cancelled',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '',
-      estArrival: '11:15',
-      estDeparture: '12:10',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'Cancelled',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '',
-      estArrival: '12:00',
-      estDeparture: '13:30',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '15.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1040',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 1 },
-  ),
-
-  // row-group
-  // 16.10.2022
-  setItemPayload(group2, {
-    actionType: 'row-group',
-    parent: true,
-    groupId: 2,
-    rules: {
-      overrides: {
-        collapse: {
-          action: {
-            componentType: MatDatagridCollapseComponent,
-            componentPosition: 'override',
-            data: {
-              item: group2,
-            },
-          },
-        },
-      },
-    },
-  }),
-
-  // row-single
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: '',
-      status: 'Boarding',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '07:50',
-      estDeparture: '08:45',
-      terminal: '2',
-      gate: 'A02',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Central Ways',
-      flightNumber: '-2415',
-      status: 'Boarding',
-      origin: 'Hamburg',
-      destination: 'Izmir',
-      duration: '3h 10min',
-      estArrival: '09:00',
-      estDeparture: '09:45',
-      terminal: '2',
-      gate: 'A01',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1932',
-      status: 'Boarding',
-      origin: 'Hamburg',
-      destination: 'Oslo',
-      duration: '1h 25min',
-      estArrival: '09:15',
-      estDeparture: '10:00',
-      terminal: '2',
-      gate: 'B08',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-8551',
-      status: 'Boarding',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:30',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: '',
-      origin: 'Hamburg',
-      destination: 'Kos',
-      duration: '3h 25min',
-      estArrival: '10:45',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: '',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:15',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1435',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '12:15',
-      terminal: '1',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1126',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Helsinki',
-      duration: '1h 50min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '16.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1044',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'C10',
-    },
-    { actionType: 'row-single', groupId: 2 },
-  ),
-
-  // row-group
-  // 17.10.2022
-  setItemPayload(group3, {
-    actionType: 'row-group',
-    parent: true,
-    groupId: 3,
-    rules: {
-      overrides: {
-        collapse: {
-          action: {
-            componentType: MatDatagridCollapseComponent,
-            componentPosition: 'override',
-            data: {
-              item: group3,
-            },
-          },
-        },
-      },
-    },
-  }),
-
-  // row-single
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '17.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 18.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '18.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 19.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '19.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 20.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '20.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 21.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '21.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 22.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '22.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-
-  // 23.10.2022
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-2422',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Dublin',
-      duration: '2h min',
-      estArrival: '07:40',
-      estDeparture: '08:30',
-      terminal: '2',
-      gate: 'C02',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-3931',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Antalya',
-      duration: '3h 55min',
-      estArrival: '08:15',
-      estDeparture: '09:10',
-      terminal: '1',
-      gate: 'A03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-0274',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Madrid',
-      duration: '2h 10min',
-      estArrival: '09:00',
-      estDeparture: '10:00',
-      terminal: '1',
-      gate: 'B04',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Comfort Flights',
-      flightNumber: 'CF-7532',
-      status: 'On Time',
-      origin: 'London',
-      destination: 'Istanbul',
-      duration: '3h 55min',
-      estArrival: '09:45',
-      estDeparture: '10:30',
-      terminal: '1',
-      gate: 'B13',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Astro Airlines',
-      flightNumber: 'AA-1264',
-      status: 'On Time',
-      origin: 'Berlin',
-      destination: 'London',
-      duration: '1h 55min',
-      estArrival: '10:30',
-      estDeparture: '11:15',
-      terminal: '2',
-      gate: 'A09',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Sky Fly',
-      flightNumber: 'SF-5323',
-      status: 'On Time',
-      origin: 'Frankfurt',
-      destination: 'Madrid',
-      duration: '2h 40min',
-      estArrival: '11:00',
-      estDeparture: '11:45',
-      terminal: '1',
-      gate: 'C03',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1632',
-      status: 'On Time',
-      origin: 'Munich',
-      destination: 'Rome',
-      duration: '1h 30min',
-      estArrival: '11:15',
-      estDeparture: '',
-      terminal: '2',
-      gate: 'A12',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Venture Airlines',
-      flightNumber: 'VA-1424',
-      status: 'On Time',
-      origin: 'Prague',
-      destination: 'Brussel',
-      duration: '1h 30min',
-      estArrival: '12:00',
-      estDeparture: '13:00',
-      terminal: '1',
-      gate: 'A07',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-  setItemPayload(
-    {
-      collapse: '',
-      date: '23.10.2022',
-      airline: 'Central Ways',
-      flightNumber: 'CW-1042',
-      status: 'On Time',
-      origin: 'Hamburg',
-      destination: 'Danzig',
-      duration: '1h 30min',
-      estArrival: '13:45',
-      estDeparture: '14:50',
-      terminal: '2',
-      gate: 'B10',
-    },
-    { actionType: 'row-single', groupId: 3 },
-  ),
-];
